third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferToGPUTexture-initialized-canvas.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./canvas-test.js"></script>
<script src="./webgpu-helpers.js"></script>

<h1>Canvas2D WebGPU Transfer: transferToGPUTexture-initialized-canvas</h1>
<p>transferToGPUTexture() creates a texture from an initialized canvas.</p>

<script>

runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

/**
 * transferToGPUTexture() should create a texture from an initialized canvas.
 */
canvasPromiseTest(async (canvas) => {
  const device = await getWebGPUDevice();

  // Paint into the canvas to ensure it has a valid image bitmap.
  const ctx = canvas.getContext('2d');
  ctx.fillStyle = "#FFFFFF";
  ctx.fillRect(0, 0, 100, 50);

  // Next, call transferToGPUTexture.
  const tex = ctx.transferToGPUTexture({device: device,
                                        label: "hello, webgpu!"});

  // Confirm that we now have a GPU texture that matches our request.
  assert_true(tex instanceof GPUTexture, 'not a GPUTexture');
  assert_equals(tex.label, 'hello, webgpu!');
  assert_equals(tex.width, canvas.width);
  assert_equals(tex.height, canvas.height);
}, 'transferToGPUTexture should create a texture from an initialized canvas.');

</script>
